+++ components/layout/index.tsx
@@ -1,0 +1,13 @@
+import { PropsWithChildren } from "react";
+import { Header } from "@/components/header";
+
+function Layout({ children }: PropsWithChildren<{}>) {
+  return (
+    <>
+      <Header />
+      {children}
+    </>
+  );
+}
+
+export default Layout;
