+++ components/profile/index.tsx
@@ -1,0 +1,20 @@
+import { Center, Heading } from "@chakra-ui/layout";
+import type { NextPage } from "next";
+import Head from "next/head";
+import React from "react";
+
+const Home: NextPage = () => {
+  return (
+    <Center h="calc(100vh - 96px)">
+      <Head>
+        <title>MetaFactory - Dashboard</title>
+        <meta name="description" content="MetaFactory Dashboard" />
+      </Head>
+      <Heading as="h1" textAlign="center">
+        Profile will be here
+      </Heading>
+    </Center>
+  );
+};
+
+export default Home;
