+++ pages/_app.tsx
@@ -1,18 +1,7 @@
-import "../styles/globals.css";
+import "@/styles/globals.css";
 import type { AppProps } from "next/app";
-import React, { PropsWithChildren } from "react";
 import { ChakraProvider } from "@chakra-ui/react";
-
-import { Header } from "../components/header";
-
-function Layout({ children }: PropsWithChildren<{}>) {
-  return (
-    <>
-      <Header />
-      {children}
-    </>
-  );
-}
+import Layout from "@/components/layout";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
