+++ pages/index.tsx
@@ -1,20 +1,3 @@
-import { Center, Heading } from "@chakra-ui/layout";
-import type { NextPage } from "next";
-import Head from "next/head";
-import React from "react";
+import Profile from '@/components/profile';
 
-const Home: NextPage = () => {
-  return (
-    <Center h="calc(100vh - 96px)">
-      <Head>
-        <title>MetaFactory - Dashboard</title>
-        <meta name="description" content="MetaFactory Dashboard" />
-      </Head>
-      <Heading as="h1" textAlign="center">
-        MF Dashboard home
-      </Heading>
-    </Center>
-  );
-};
-
-export default Home;
+export default Profile;
